Final code changes for 0.2.1 release.

+++ SlimTuneUI/DummyStorageEngine.cs
@@ -52,7 +52,7 @@ namespace SlimTuneUI
 		{
 		}
 
-		public void PerfCounter(int counterId, long time, long value)
+		public void PerfCounter(int counterId, long time, double value)
 		{
 		}
