+++ SlimTuneUI/SqlServerCompactEngine.cs
@@ -168,7 +168,7 @@ namespace SlimTuneUI
 		{
 		}
 
-		public override void PerfCounter(int counterId, long time, long value)
+		public override void PerfCounter(int counterId, long time, double value)
 		{
 		}
