+++ SlimTuneUI/UICore/StorageEngineBase.cs
@@ -86,7 +86,7 @@ namespace UICore
 		public abstract void MapClass(ClassInfo classInfo);
 		public abstract void UpdateThread(int threadId, bool? alive, string name);
 		public abstract void CounterName(int counterId, string name);
-		public abstract void PerfCounter(int counterId, long time, long value);
+		public abstract void PerfCounter(int counterId, long time, double value);
 		public abstract void Flush();
 		public abstract void Save(string file);
 		public abstract void Snapshot(string name);
